+++ src/uniprotkb/types/searchTypes.ts
@@ -1,21 +1,3 @@
-export enum Namespace {
-  uniprotkb = 'uniprotkb',
-  uniref = 'uniref',
-  uniparc = 'uniparc',
-  proteomes = 'proteomes',
-  publications = 'publications',
-  keywords = 'keywords',
-}
-
-export const NamespaceLabels = {
-  [Namespace.uniprotkb]: 'UniProtKB',
-  [Namespace.uniref]: 'UniRef',
-  [Namespace.uniparc]: 'UniParc',
-  [Namespace.proteomes]: 'Proteomes',
-  [Namespace.publications]: 'Publications',
-  [Namespace.keywords]: 'Keywords',
-};
-
 export enum Operator {
   AND = 'AND',
   OR = 'OR',
